Snakes and Ladders game
This is an implementation of the game Snakes and Ladders created with  CSS / HTML / JavaScript.
Two players play (the first player that will start is the one that brings the biggest number on a dice roll) while the players are represented by their unique pawns for one player each. The goal is to reach exactly the end of the board, not more or less but at the exact end (position 90). 
The game is themed as a University Snakes and Ladders so you win when you get the degree!